+++ src/route/AuthenticatedRoute/AuthenticatedRoute.module.css
@@ -1,0 +1,3 @@
+.App {
+  display: flex;
+}
